HLEP

+++ src/server/ecs/EcsComponents.h
@@ -6,12 +6,12 @@ namespace lol_at_home_server {
 
 struct PendingDamage {
   entt::entity source;  // Who dealt the damage
-  double amount;
+  float amount;
 };
 
 struct PendingHeal {
   entt::entity source;  // Who healed
-  double amount;
+  float amount;
 };
 
 }  // namespace lol_at_home_server
